src/test/java/example/HardToTestExampleControllerTest.java
package example;

import example.person.Person;
import example.person.PersonRepository;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.context.junit4.SpringRunner;

import java.util.Optional;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.core.Is.is;
import static org.mockito.BDDMockito.given;

@SpringBootTest
@RunWith(SpringRunner.class)
public class HardToTestExampleControllerTest {

    @Autowired
    private HardToTestExampleController subject;

    @MockBean
    private PersonRepository personRepository;

    @Test
    public void shouldReturnFullNameOfAPerson() throws Exception {
        Person peter = new Person("Peter", "Pan");
        given(personRepository.findByLastName("Pan")).willReturn(Optional.of(peter));

        var greeting = subject.hello("Pan");

        assertThat(greeting, is("Hello Peter Pan!"));
    }
}